refactor(distiller): split issue-draft into focused modules
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
