+++ packages/distillers/issue-draft/src/index.ts
@@ -1,248 +1,12 @@
-import type { DistillResultDraft, DistillerFactory, DistillerRunInput, SessionArtifact } from "@loamlog/core";
-import { createEvidence, defineDistiller } from "@loamlog/distiller-sdk";
+import type { DistillResultDraft, DistillerFactory, DistillerRunInput } from "@loamlog/core";
+import { defineDistiller } from "@loamlog/distiller-sdk";
 
-const DISTILLER_ID = "@loamlog/distiller-issue-draft";
-const MAX_MESSAGE_CHARS = 1200;
-const SUPPORTED_ISSUE_KINDS = new Set(["bug", "feature", "docs", "refactor", "chore"]);
-const SYSTEM_PROMPT = [
-  "You extract exactly one strong GitHub issue draft from an AI coding session.",
-  "Return a JSON array only.",
-  "Each item must include: title, summary, background, problem, proposed_solution, acceptance_criteria, confidence, evidence_refs.",
-  "Optional fields: issue_kind, labels.",
-  "Each evidence_refs item must include message_id and excerpt.",
-  "Prefer no result over weakly supported results.",
-].join("\n");
-
-type IssueKind = "bug" | "feature" | "docs" | "refactor" | "chore";
-
-interface IssueDraftPayload {
-  title: string;
-  issue_kind?: IssueKind;
-  labels?: string[];
-}
-
-interface LlmEvidenceRef {
-  message_id: string;
-  excerpt: string;
-}
-
-interface RawEvidenceRef {
-  message_id?: unknown;
-  excerpt?: unknown;
-}
-
-interface LlmIssueDraft {
-  title: string;
-  summary: string;
-  background: string;
-  problem: string;
-  proposed_solution: string;
-  acceptance_criteria: string[];
-  confidence?: number;
-  issue_kind?: string;
-  labels?: string[];
-  evidence_refs?: LlmEvidenceRef[];
-}
-
-function estimateTokens(content: string): number {
-  return Math.max(1, Math.ceil(content.length / 4));
-}
-
-function buildPrompt(artifact: SessionArtifact): string {
-  const chunks = artifact.messages.map((message: SessionArtifact["messages"][number]) => {
-    const text = (message.content ?? "").slice(0, MAX_MESSAGE_CHARS);
-    return `[${message.id}] (${message.role}) ${text}`;
-  });
-
-  return [
-    `session_id: ${artifact.meta.session_id}`,
-    "messages:",
-    ...chunks,
-    "",
-    "Output format:",
-    '[{"title":"...","summary":"...","background":"...","problem":"...","proposed_solution":"...","acceptance_criteria":["..."],"confidence":0.0,"issue_kind":"feature","labels":["triage"],"evidence_refs":[{"message_id":"...","excerpt":"..."}]}]',
-  ].join("\n");
-}
-
-function extractJsonPayload(content: string): string {
-  const trimmed = content.trim();
-  if (trimmed.startsWith("[") || trimmed.startsWith("{")) {
-    return trimmed;
-  }
-
-  const fenced = trimmed.match(/```(?:json)?\s*([\s\S]*?)\s*```/i);
-  if (fenced && fenced[1]) {
-    return fenced[1].trim();
-  }
-
-  return trimmed;
-}
-
-function isStringArray(value: unknown): value is string[] {
-  return Array.isArray(value) && value.every((item: unknown) => typeof item === "string");
-}
-
-function normalizeIssueKind(value: unknown): IssueKind | undefined {
-  if (typeof value !== "string") {
-    return undefined;
-  }
-
-  return SUPPORTED_ISSUE_KINDS.has(value) ? (value as IssueKind) : undefined;
-}
-
-function normalizeLabels(value: unknown): string[] | undefined {
-  if (!isStringArray(value)) {
-    return undefined;
-  }
-
-  const labels = value.map((label: string) => label.trim()).filter((label: string) => label.length > 0);
-  return labels.length > 0 ? Array.from(new Set(labels)) : undefined;
-}
-
-function normalizeText(value: string): string {
-  return value.trim();
-}
-
-function parseIssueDrafts(content: string): LlmIssueDraft[] {
-  const json = extractJsonPayload(content);
-  const parsed = JSON.parse(json) as unknown;
-  if (!Array.isArray(parsed)) {
-    return [];
-  }
-
-  return parsed.filter((item: unknown): item is LlmIssueDraft => {
-    if (!item || typeof item !== "object") {
-      return false;
-    }
-
-    const candidate = item as Record<string, unknown>;
-    return (
-      typeof candidate.title === "string" &&
-      normalizeText(candidate.title).length > 0 &&
-      typeof candidate.summary === "string" &&
-      normalizeText(candidate.summary).length > 0 &&
-      typeof candidate.background === "string" &&
-      normalizeText(candidate.background).length > 0 &&
-      typeof candidate.problem === "string" &&
-      normalizeText(candidate.problem).length > 0 &&
-      typeof candidate.proposed_solution === "string" &&
-      normalizeText(candidate.proposed_solution).length > 0 &&
-      isStringArray(candidate.acceptance_criteria)
-    );
-  });
-}
-
-function findMessage(artifact: SessionArtifact, messageId: string): SessionArtifact["messages"][number] | undefined {
-  return artifact.messages.find((message: SessionArtifact["messages"][number]) => message.id === messageId);
-}
-
-function buildEvidence(artifact: SessionArtifact, refs: LlmEvidenceRef[] | undefined) {
-  const seen = new Set<string>();
-  const evidence: DistillResultDraft["evidence"] = [];
-
-  for (const rawRef of refs ?? []) {
-    if (!rawRef || typeof rawRef !== "object") {
-      continue;
-    }
-
-    const ref = rawRef as RawEvidenceRef;
-    if (typeof ref.message_id !== "string" || typeof ref.excerpt !== "string") {
-      continue;
-    }
-
-    const messageId = ref.message_id.trim();
-    const excerpt = ref.excerpt.trim();
-    if (messageId.length === 0 || excerpt.length === 0) {
-      continue;
-    }
-
-    const message = findMessage(artifact, messageId);
-    if (!message) {
-      continue;
-    }
-
-    const key = `${messageId}:${excerpt}`;
-    if (seen.has(key)) {
-      continue;
-    }
-
-    seen.add(key);
-    evidence.push(createEvidence(artifact, message, excerpt));
-  }
-
-  return evidence;
-}
-
-function renderMarkdown(issue: LlmIssueDraft, evidence: DistillResultDraft["evidence"]): string {
-  const acceptanceCriteria = issue.acceptance_criteria.map((item: string) => `- ${item}`).join("\n");
-  const evidenceLines = evidence
-    .map((item: DistillResultDraft["evidence"][number]) => `- \`${item.message_id}\`: ${item.excerpt}`)
-    .join("\n");
-  const labelLine = normalizeLabels(issue.labels);
-  const issueKind = normalizeIssueKind(issue.issue_kind);
-
-  return [
-    issueKind ? `Type: ${issueKind}` : undefined,
-    labelLine ? `Labels: ${labelLine.join(", ")}` : undefined,
-    "## Background",
-    issue.background,
-    "",
-    "## Problem",
-    issue.problem,
-    "",
-    "## Proposed Solution",
-    issue.proposed_solution,
-    "",
-    "## Acceptance Criteria",
-    acceptanceCriteria,
-    "",
-    "## Evidence",
-    evidenceLines,
-  ]
-    .filter((line: string | undefined): line is string => line !== undefined)
-    .join("\n");
-}
-
-function toTags(issue: LlmIssueDraft): string[] {
-  const tags = ["issue-draft"];
-  const issueKind = normalizeIssueKind(issue.issue_kind);
-  if (issueKind) {
-    tags.push(issueKind);
-  }
-
-  const labels = normalizeLabels(issue.labels);
-  if (labels) {
-    tags.push(...labels);
-  }
-
-  return Array.from(new Set(tags));
-}
-
-function selectBestCandidate(candidates: LlmIssueDraft[], artifact: SessionArtifact): {
-  issue: LlmIssueDraft;
-  evidence: DistillResultDraft["evidence"];
-} | null {
-  const scored = candidates
-    .map((issue: LlmIssueDraft) => ({ issue, evidence: buildEvidence(artifact, issue.evidence_refs) }))
-    .filter((candidate) => candidate.evidence.length > 0)
-    .sort((left, right) => {
-      const confidenceDelta =
-        (typeof right.issue.confidence === "number" ? right.issue.confidence : 0.5) -
-        (typeof left.issue.confidence === "number" ? left.issue.confidence : 0.5);
-      if (confidenceDelta !== 0) {
-        return confidenceDelta;
-      }
-
-      const evidenceDelta = right.evidence.length - left.evidence.length;
-      if (evidenceDelta !== 0) {
-        return evidenceDelta;
-      }
-
-      return normalizeText(left.issue.title).localeCompare(normalizeText(right.issue.title));
-    });
-
-  return scored[0] ?? null;
-}
+import { DISTILLER_ID, SYSTEM_PROMPT } from "./constants.js";
+import { normalizeIssueKind, normalizeLabels, normalizeText, parseIssueDrafts } from "./parse.js";
+import { buildPrompt, estimateTokens } from "./prompt.js";
+import { renderMarkdown, toTags } from "./render.js";
+import { selectBestCandidate } from "./select.js";
+import type { IssueDraftPayload } from "./types.js";
 
 const factory: DistillerFactory = () =>
   defineDistiller<IssueDraftPayload>({
